Authorization on services

+++ src/main/java/com/github/italomded/recipesapi/service/ImageService.java
@@ -1,10 +1,12 @@
 package com.github.italomded.recipesapi.service;
 
+import com.github.italomded.recipesapi.domain.ApplicationUser;
 import com.github.italomded.recipesapi.domain.Image;
 import com.github.italomded.recipesapi.domain.Recipe;
 import com.github.italomded.recipesapi.dto.form.ImageForm;
 import com.github.italomded.recipesapi.repository.ImageRepository;
 import com.github.italomded.recipesapi.repository.RecipeRepository;
+import com.github.italomded.recipesapi.service.exception.BusinessRuleException;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -27,18 +29,20 @@ public class ImageService {
     }
 
     @Transactional
-    public Image editImage(Long imageID, ImageForm form) {
-        // TODO: verify if request user is the author of the recipe
+    public Image editImage(Long imageID, ImageForm form, ApplicationUser userAuthor) {
         Image image = imageRepository.getReferenceById(imageID);
+        RecipeService.verifyIfIsTheAuthor(image.getRecipe(), userAuthor);
+
         image.setImageBytes(form.bytes());
         imageRepository.save(image);
         return image;
     }
 
     @Transactional
-    public Image createImage(Long recipeID, ImageForm form) {
-        // TODO: verify if request user is the author of the recipe
+    public Image createImage(Long recipeID, ImageForm form, ApplicationUser userAuthor) {
         Recipe recipe = recipeRepository.getReferenceById(recipeID);
+        RecipeService.verifyIfIsTheAuthor(recipe, userAuthor);
+
         Image image = new Image(form.bytes(), recipe);
         recipe.addImage(image);
 
@@ -48,11 +52,18 @@ public class ImageService {
     }
 
     @Transactional
-    public void deleteImage(Long imageID) {
-        // TODO: verify if request user is the author of the recipe
+    public void deleteImage(Long imageID, ApplicationUser userAuthor) {
         Image image = imageRepository.getReferenceById(imageID);
-        image.getRecipe().removeImage(image);
-        recipeRepository.save(image.getRecipe());
-        imageRepository.delete(image);
+        RecipeService.verifyIfIsTheAuthor(image.getRecipe(), userAuthor);
+
+        boolean removed = image.getRecipe().removeImage(image);
+        if (removed) {
+            recipeRepository.save(image.getRecipe());
+            imageRepository.delete(image);
+        } else {
+            throw new BusinessRuleException(
+                    Recipe.class,
+                    "can't delete recipe image because the image list should have at least 1 image or the image does not belong in the recipe");
+        }
     }
 }
